Add TCP port scanner with banner grabbing functionality
This commit introduces a Python script that performs TCP port scanning on a user-specified IP address. It iterates through ports in the range of 1 to a user-defined upper limit, attempting to establish a connection to each. Upon a successful connection, it tries to retrieve the service banner, providing additional context about the service running on the open port.